requirement.txt
Django==1.9.1
-e git+https://timmyomahony@github.com/timmyomahony/django-pagedown.git@c7d158dfeca312fa4eecbbab1601f472d645ccf4#egg=django_pagedown-master
django-taggit==0.18.0
ecdsa==0.13
Fabric==1.10.2
gunicorn==19.4.5
Markdown==2.6.5
paramiko==1.16.0
pycrypto==2.6.1
Pygments==2.1
pytz==2015.7
South==1.0.2
wheel==0.24.0
xmltodict==0.9.2
xmlutils==1.1
